GPareto-class(distrEx)	R Documentation

Generalized Pareto distribution

Description

[borrowed from evd]:

The (Three-parameter) generalized Pareto distribution with parameter loc= a, scale = b, shape = c has density:

f(x) = frac{1}{b} (1+c z)^(-1/c - 1), quad z = frac{x-a}{c}

for x > a ( c >=q 0) and a <=q x <=q a - b/c(c < 0).

Objects from the Class

Objects can be created by calls of the form new("GPareto", loc, scale,shape). More frequently they are created via the generating function GPareto.

Slots

img:

Object of class "Reals".

param:

Object of class "GParetoParameter".

r:

rgpd

d:

dgpd

p:

pgpd, but vectorized and with special treatment of arguments lower.tail and log.p

q:

qgpd, but vectorized and with special treatment of arguments lower.tail and log.p

gaps:

(numeric) matrix or NULL

.withArith:

logical: used internally to issue warnings as to interpretation of arithmetics

.withSim:

logical: used internally to issue warnings as to accuracy

.logExact:

logical: used internally to flag the case where there are explicit formulae for the log version of density, cdf, and quantile function

.lowerExact:

logical: used internally to flag the case where there are explicit formulae for the lower tail version of cdf and quantile function

Extends

Class "AbscontDistribution", directly.
Class "UnivariateDistribution", by class "AbscontDistribution".
Class "Distribution", by class "AbscontDistribution".

Methods

initialize

signature(.Object = "GPareto"): initialize method.

shape

signature(object = "GPareto"): wrapped access method for slot shape of slot param.

loc

signature(object = "GPareto"): wrapped access method for slot loc of slot param.

location

signature(object = "GPareto"): alias to loc, to support argument naming of package VGAM.

scale

signature(x = "GPareto"): wrapped access method for slot scale of slot param.

shape<-

signature(object = "GPareto"): wrapped replace method for slot shape of slot param.

loc<-

signature(object = "GPareto"): wrapped replace method for slot loc of slot param.

location<-

signature(object = "GPareto"): alias to loc<-, to support argument naming of package VGAM.

scale<-

signature(x = "GPareto"): wrapped replace method for slot scale of slot param.

+

signature(e1 = "GPareto", e2 = "numeric"): exact method for this transformation — stays within this class.

*

signature(e1 = "GPareto", e2 = "numeric"): exact method for this transformation — stays within this class if e2>0.

E

signature(object = "GPareto", fun = "missing", cond = "missing"): exact evaluation using explicit expressions.

var

signature(signature(x = "GPareto"): exact evaluation using explicit expressions.

median

signature(signature(x = "GPareto"): exact evaluation using explicit expressions.

IQR

signature(signature(x = "GPareto"): exact evaluation using explicit expressions.

skewness

signature(signature(x = "GPareto"): exact evaluation using explicit expressions.

kurtosis

signature(signature(x = "GPareto"): exact evaluation using explicit expressions.

Note

This class is based on the code provided by the package evd by A. G. Stephenson.

Author(s)

Nataliya Horbenko Nataliya.Horbenko@itwm.fraunhofer.de

References

Pickands, J. (1975) Statistical inference using extreme order statistics. _Annals of Statistics_, *3*, 119-131.

See Also

dgpd, AbscontDistribution-class

Examples


(P1 <- new("GPareto", loc = 0, scale = 1,shape = 0))
plot(P1)
shape(P1)
loc(P1)
scale(P1) <- 4
loc(P1) <- 2
plot(P1)

[Package distrEx version 2.2 Index]